frontend: serve /repo with HTTP over dl.flathub.org
Fastly only permits "streaming misses" (sending the file to the client while it
is still being retreived from the origin) when the origin is unencrypted HTTP.
We were seeing significant origin latency which I believe is due to large files
taking a while to be fetched into the cache.

We've configured Fastly to use streaming misses on data files (.filez and delta
parts) which are otherwise hashed/verified, so we're OK with them being sourced
from the origin without encryption in order to get the speed increase, so we
serve /repo over HTTP (without directing to HTTPS) on the dl.flathub.org host
now.