frontend: never cache summary{,.sig}
We had a situation where front-hex could not reach repo due to a Mythic Beasts
routing problem, and so it had cached a stale summary and summary.sig. However,
Fastly was able to reach it and was including it in the round robin, meaning it
was non-deterministically caching a mismatched pair of summary/.sig files. Now
that we instruct Fastly to soft-purge the summary/.sig files and serve stale
ones on error, remove the possibility that a mismatched pair is fetched through
round-robining the intermediate servers by simply disabling the cache for these
files.