frontend: Proxy flatpak.org
Proxy the flatpak.org site from front.flathub.org like all of the
flathub.org sites. Primarily this is for consistency so that the
frontend servers are truly the frontend for all websites. However, it
clearly will provided a better experience for users.

The nginx configuration is mostly copied from webroot. Note that
default.d snippets are not used because that would interfere with
default including all of the snippets since they currently represent
different flathub paths.

Fixes flathub/flathub#707.