Add role for flathub-stats collection
Create a role for collecting stats on flathub. This was previously
running on front.flathub.org but will now be running on the
stats.flathub.org machine in packet. There are 2 components:

* An rsyslog instance receiving a syslog stream from the Fastly CDN
  providing dl.flathub.org. A configuration is added to log the streams
  to /var/cache/flathub-stats/logs. A custom firewalld zone is added so
  that the service is only available to Fastly's IP addresses.

* A script running from the flathub-stats repo that analyzes the log
  messages and outputs a JSON summary. The script is run from a
  logrotate configuration and needs some SELinux transitions to run
  correctly. A custom SELinux module is installed to provide the
  transitions and the file context rules. The summaries are output to
  /srv/flathub-stats/stats on a persistent volume.

The CDN configuration has already been updated to stream the logs to
this machine in addition to front.