dehydrated: Support certsync to remote host
Extend the deploy_cert hook so that it can send acquired certificates to
remote hosts. This is used to send acquired certificates to the
frontends. The hook here is basically the same as what's on front except
that the domains are looped over instead of trying to construct a case
statement from jinja.