dehydrated: Create root SSH key if necessary when syncing
If there are certificate domains to sync, ensure that root has an SSH
key present. A 4096 bit RSA key is generated if one doesn't exist.
Obviously, a generated SSH key isn't that helpful since the remote side
needs to have the associated public key, but this is better than nothing
in the absence of useful secret persistence.