frontend: Refactor certsync SSH authorized keys
Instead of inlining the authorized SSH keys for certsync, move
root@front's SSH key to a file and list the allowed users in the new
frontend_certsync_ssh_users variable. This will make it cleaner to add
new certsync initiators. The key file is added to the top level as the
keys could be reused in other contexts.