base: add support for Mythic Beasts backup service
mythic-backup runs daily on cron and basically rsyncs all of the filesystems to
a host provided by Mythic, storing the permissions as xattrs. It also supports
dumping postgres databases, which we key off the new postgres group. Backups
exclude large/generated data sets as we have 50GB of space, and are disabled on
the frontend servers as they have no relevant persistent state and can be
re-provisioned whenever.